chore: removed extra wrappers for home page and rename health endpoint
