+++ src/pages/api/health.ts
@@ -1,8 +1,5 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
-export default function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<any>,
-) {
+export default function handler(_: NextApiRequest, res: NextApiResponse<any>) {
   res.status(200).send('Healthy');
 }
