--- src/pages/api/home/index.ts
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from './home';
